SF-162: colocate Anthropic bootstrap tests
Move Claude Code bootstrap coverage into the Anthropic test package so provider-specific behavior lives with the provider while generic startup tests cover only the plugin hook contract.

Asana: https://app.asana.com/0/1213628819033917/1214568319770655/f
